+++ lib/ui/homePages/diyPage.dart
@@ -1,0 +1,11 @@
+import 'package:flutter/material.dart';
+
+class MyDiy extends StatelessWidget {
+  const MyDiy({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return 
+    Center(child: Text("DIY", style: TextStyle(fontSize: 60),));
+  }
+}
