+++ lib/ui/homePages/historialPage.dart
@@ -1,0 +1,12 @@
+import 'package:flutter/material.dart';
+
+
+class MyHistorial extends StatelessWidget {
+  const MyHistorial({super.key});
+
+  @override
+  Widget build(BuildContext context) {
+    return 
+    Center(child: Text("Historial", style: TextStyle(fontSize: 60),));
+  }
+}
