+++ lib/ui/home_page_user.dart
@@ -1,7 +1,12 @@
 import 'package:flutter/material.dart';
-import 'package:we_fix_it/ui/report_page.dart'; 
+import 'package:we_fix_it/ui/homePages/diyPage.dart';
+import 'package:we_fix_it/ui/homePages/historialPage.dart';
+import 'package:we_fix_it/ui/homePages/toolsPage.dart';
 import 'package:we_fix_it/ui/widgets/widgetlogin.dart'; 
-import 'package:firebase_auth/firebase_auth.dart';
+import 'package:we_fix_it/ui/homePages/profilePage.dart';
+import 'package:we_fix_it/ui/homePages/homePage.dart';
+import 'package:we_fix_it/ui/homePages/historialPage.dart';
+import 'package:we_fix_it/ui/homePages/toolsPage.dart';
 
 class HomePage extends StatefulWidget {
   const HomePage({super.key});
@@ -14,10 +19,10 @@ class HomePage extends StatefulWidget {
 class _HomePageState extends State<HomePage> {
   int currentIndex = 2;
   final pantallas =[
-    Center(child: Text("Historial", style: TextStyle(fontSize: 60),)),
-    Center(child: Text("DIY", style: TextStyle(fontSize: 60),)),
+    MyHistorial(),
+    MyDiy(),
     MyInicio(),
-    Center(child: Text("Herramientas", style: TextStyle(fontSize: 60),)),
+    MyTools(),
     MyLogOut(),
   ];
   @override
@@ -70,98 +75,4 @@ class _HomePageState extends State<HomePage> {
       ),
     );
   }
-}
-class MyInicio extends StatefulWidget {
-  const MyInicio({super.key});
-
-  @override
-  State<MyInicio> createState() => _MyInicioState();
-}
-
-class _MyInicioState extends State<MyInicio> {
-  @override
-  Widget build(BuildContext context) => Scaffold(
-    body: Stack(
-      children: <Widget> [Align(
-        alignment: Alignment.topRight,
-        child: Padding(
-          padding: EdgeInsets.only(right: 50),
-          child: ElevatedButton(
-                style: ElevatedButton.styleFrom(
-                  backgroundColor: Color.fromARGB(255, 232, 232, 232),
-                  foregroundColor: Colors.black,
-                  shadowColor: Colors.black,
-                  elevation: 3,
-                  shape: RoundedRectangleBorder(
-                      borderRadius: BorderRadius.circular(5)),
-                  minimumSize: Size(300, 40), 
-                ),
-                onPressed: () => Navigator.push(
-                context,
-                MaterialPageRoute(builder: (context) => MyReportPage(title: "Reportar Problema")),
-                ),
-                child: Text('¡Reportanos un problema!', style: TextStyle(fontSize: 20),),
-              ),
-        ),
-      ),
-          Align(
-            alignment: Alignment.topRight,
-            child: IconButton(onPressed: () => Navigator.push(
-              context,
-              MaterialPageRoute(builder: (context) => MyReportPage(title: "Reportar Problema")),
-              ), 
-              icon: Icon(Icons.email),
-              iconSize: 35,
-              color: Colors.black,
-              ),
-          )
-          ]
-    )
-    );
-}
-class MyLogOut extends StatefulWidget {
-  const MyLogOut({super.key});
-
-  @override
-  State<MyLogOut> createState() => _MyLogOutState();
-}
-
-class _MyLogOutState extends State<MyLogOut> {
-  @override
-  Widget build(BuildContext context) => Scaffold(
-      body: Center(
-        child: Container( 
-          height: 200,
-          width: 350,
-          decoration: BoxDecoration(
-            borderRadius: BorderRadius.circular(10),
-            color: Color.fromARGB(255, 232, 232, 232),
-          ),
-          child: Column(
-            children: [
-              const SizedBox(height: 25),
-              Text("¿Esta seguro que desea cerrar sesión?", style: TextStyle(fontSize: 18),),
-
-              const SizedBox(height: 25),
-
-              ElevatedButton(
-              onPressed: () async{
-                await FirebaseAuth.instance.signOut();
-              },
-              style: ElevatedButton.styleFrom(
-                    backgroundColor: Color.fromARGB(255, 254, 182, 15),
-                    foregroundColor: Colors.black,
-                    shadowColor: Colors.black,
-                    elevation: 3,
-                    shape: RoundedRectangleBorder(
-                        borderRadius: BorderRadius.circular(5)),
-                    minimumSize: Size(250, 70), 
-                  ),
-              child: const Text("Cerrar Sesión", style: TextStyle(fontSize: 24),),
-            ),
-            ]          
-          ),
-        ),
-      ),
-    );
 }
